requirements.txt
mistralai==1.6.0
langchain
langchain-openai==0.3.12
langchain-chroma==0.2.2
langchain-community==0.3.21
fastembed==0.6.0
python-dotenv==1.1.0
